+++ components/service/base_obj.h
@@ -35,22 +35,22 @@ struct BaseObj
     /* 设置波形对应通道 name*/
     void (*setChName)(struct BaseObj *that, uint16_t ch, char*name,uint32_t len);
     /* 对应通道添加波形数据*/
-    void (*addWave)(struct BaseObj *that, uint16_t ch, double data);
+    void (*addWave)(struct BaseObj *that, uint16_t ch, float data);
 
     /* 获取时间cnt*/
     uint64_t (*getTimeCnt)(struct BaseObj *that);
     /* 获取时间ms*/
-    double (*getTimeMs)(struct BaseObj *that);
+    float (*getTimeMs)(struct BaseObj *that);
     /* 获取时间us*/
-    double (*getTimeUs)(struct BaseObj *that);
+    float (*getTimeUs)(struct BaseObj *that);
     /* 获取时间S*/
-    double (*getTimeS)(struct BaseObj *that);
-    double (*getTimeHs)(struct BaseObj *that);
+    float (*getTimeS)(struct BaseObj *that);
+    float (*getTimeHs)(struct BaseObj *that);
 
     /* 全局波形打印*/
     void (*globalPlot)(struct BaseObj *that);
     /* 全局时间设置补偿*/
-    void (*globalInitTime)(struct BaseObj *that,double hs);
+    void (*globalInitTime)(struct BaseObj *that,float hs);
 
     /* 周期循环仿真全局时间步进*/
     void (*globalTimeStep)(struct BaseObj *that);
@@ -61,14 +61,14 @@ struct BaseObj
         struct \
         {\
             void (*setChName)(struct BaseObj *that, uint16_t ch, char*name,uint32_t len);\
-            void (*addWave)(struct BaseObj *that, uint16_t ch, double data);\
+            void (*addWave)(struct BaseObj *that, uint16_t ch, float data);\
             uint64_t (*getTimeCnt)(struct BaseObj *that);\
-            double (*getTimeMs)(struct BaseObj *that);\
-            double (*getTimeUs)(struct BaseObj *that);\
-            double (*getTimeS)(struct BaseObj *that);\
-            double (*getTimeHs)(struct BaseObj *that);\
+            float (*getTimeMs)(struct BaseObj *that);\
+            float (*getTimeUs)(struct BaseObj *that);\
+            float (*getTimeS)(struct BaseObj *that);\
+            float (*getTimeHs)(struct BaseObj *that);\
             void (*globalPlot)(struct BaseObj *that);\
-            void (*globalInitTime)(struct BaseObj *that,double hs);\
+            void (*globalInitTime)(struct BaseObj *that,float hs);\
             void (*globalTimeStep)(struct BaseObj *that);\
         };\
 }
